requirements.txt
# Flask==2.1.0
# Werkzeug==2.0.2
# pymysql==1.0.2
Blinker       ==  1.8.2
Click         ==  8.1.7
colorama      ==  0.4.6
distlib       ==  0.3.8
filelock      ==  3.14.0
Flask         ==  3.0.3
itsdangerous  ==  2.2.0
Jinja2        ==  3.1.4
MarkupSafe    ==  2.1.5
platformdirs  ==  4.2.1
virtualenv    ==  20.26.1
Werkzeug      ==  3.0.3
cryptography  ==  42.0.7
mysql-connector-python == 8.4.0